+++ src/app/handle_rep.rs
@@ -1,9 +1,9 @@
+use crate::app::reputation_message::ReputationMessage;
 use crate::objects::messages::ParseMessage;
+use crate::objects::user::generate_display_name;
 use crate::services::config::settings::Settings;
 use crate::services::config::triggers::TriggerType;
 use crate::services::data::reputations::Reputations;
-
-use super::message_data::MessageData;
 
 mod reputation_calculator;
 
@@ -15,37 +15,40 @@ pub struct HandledReputation {
 }
 
 impl HandledReputation {
-    pub fn handle_rep(data: &MessageData) -> Option<HandledReputation> {
-        if !data.get_valid() {
+    pub fn handle_rep(reputation_message: &ReputationMessage) -> Option<HandledReputation> {
+        if !reputation_message.is_valid {
             return None;
         }
 
         let operation: String;
 
         let reputations = Reputations::load();
-        let (reciv_reputation, reputations) = reputation_calculator::calculate_reputation(&data, reputations);
+        let (reciv_reputation, trigger, reputations) =
+            reputation_calculator::calculate_reputation(&reputation_message, reputations);
         let _result = Reputations::save(reputations);
 
-        match data.get_trigger_type() {
+        match trigger {
             TriggerType::None => return None,
             TriggerType::Positive => operation = "increased".to_string(),
             TriggerType::Negative => operation = "decreased".to_string(),
         }
 
-        let reciv_username = data
-            .get_rep_reciv_user_name()
+        let reciv_username = reputation_message
+            .rep_reciv
             .as_ref()
-            .unwrap_or(&"Unknown".to_string())
+            .map(|user| generate_display_name(&user))
+            .unwrap_or("Unknown".to_string())
             .clone();
-        let giver_username = data
-            .get_rep_giver_user_name()
+        let giver_username = reputation_message
+            .rep_giver
             .as_ref()
-            .unwrap_or(&"Unknown".to_string())
+            .map(|user| generate_display_name(&user))
+            .unwrap_or("Unknown".to_string())
             .clone();
 
         let settings = Settings::load();
         if settings.save_history {
-            reputation_calculator::save_reputation_to_history(data);
+            reputation_calculator::save_reputation_to_history(reputation_message, trigger);
         }
 
         Some(HandledReputation {