+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Simple migrations system
+  - migration to new version of triggers
+### Changed
+- Reputation triggers
+  - change structure
+  - triggers are now separated by chat_id
+  - triggers now support `is_wildcard` option
+- general refactoring
+
+### Fixed
+- bug when first reputation is added in chat it was 0 not 1/-1
+
+
 ## [0.2.0] - 2022-06-13
 
 ### Added